requirements.txt
absl-py==1.3.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
faiss-cpu==1.7.2
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
gym==0.26.2
gym-notices==0.0.8
idna==3.4
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
oauthlib==3.2.2
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyboolet==1.0.1
pybullet==3.2.5
requests==2.28.1
requests-oauthlib==1.3.1
robot-gym @ git+https://github.com/TimSchneider42/robot-gym.git@d745c9018c5acc872bedfa4589c7b5bda3c296b1
rsa==4.9
scipy==1.9.2
sisyphus-env @ git+https://github.com/TimSchneider42/sisyphus-env@237177ef112de17265cef81651393385cdaac77d
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
transformation3d==1.0.1
trimesh==3.7.13
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
